Fixed bugs
- Fixed compile type error
- Removed code causing test exe to crash without assets

+++ src/dsdl/mixer/soundbank.d
@@ -24,7 +24,7 @@ class SoundBank : ResourceBank!SoundChunk {
             return;
         }
         foreach (ref string line; lns) {
-            long indexOfSpace = indexOfAny(line, " \t");
+            uint indexOfSpace = indexOfAny(line, " \t");
             if (indexOfSpace == -1 || indexOfSpace == line.length)
                 throw new DataFormatException("Unable to parse file \"" ~ chunkListFile ~ "\"");
             string chunkName = line[0..indexOfSpace];
